requirements.txt
Django==3.2.9
djangorestframework==3.12.4
pillow~=8.1.0
python-decouple
dj-database-url
dj-static
djangorestframework-simplejwt==4.4.0
djoser
django-cors-headers
PyJWT==1.7.1
python-dateutil==2.8.2